Remove features from config response, add to webhook payload
