+++ packages/back-end/src/jobs/webhooks.ts
@@ -3,6 +3,7 @@ import { WebhookModel } from "../models/WebhookModel";
 import { createHmac } from "crypto";
 import fetch from "node-fetch";
 import { getExperimentOverrides } from "../services/organizations";
+import { getFeatureDefinitions } from "../services/features";
 
 const WEBHOOK_JOB_NAME = "fireWebhook";
 type WebhookJob = Job<{
@@ -26,9 +27,11 @@ export default function (ag: Agenda) {
     if (!webhook) return;
 
     const overrides = await getExperimentOverrides(webhook.organization);
+    const features = await getFeatureDefinitions(webhook.organization);
     const payload = JSON.stringify({
       timestamp: Math.floor(Date.now() / 1000),
       overrides,
+      features,
     });
 
     const signature = createHmac("sha256", webhook.signingKey)
